# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Qwen-Valencia API
  description: |
    API Enterprise-Level para Qwen-Valencia - Sistema de ejecución de modelos de lenguaje
    con soporte para Groq API y Ollama local, incluyendo fallback inteligente, circuit breakers,
    rate limiting y manejo de errores estandarizado.
  version: 1.0.0
  contact:
    name: Qwen-Valencia Support
    url: https://github.com/GUESTVALENCIA/Qwen-Valencia
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:6003
    description: Groq API Server (desarrollo)
  - url: http://localhost:6002
    description: Ollama MCP Server (desarrollo)
  - url: http://localhost:6000
    description: MCP Universal Server (desarrollo)
...
paths:
...
  /api/v1/groq/chat:
    post:
      tags:
        - Groq API
      summary: Chat con modelo Groq
      description: Envía un mensaje a un modelo Groq y recibe la respuesta
      operationId: groqChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Respuesta exitosa
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Límite de requests por ventana
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Requests restantes en la ventana actual
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
              description: Tiempo de reset de la ventana
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Request inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API Key inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit excedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ChatRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          description: Nombre del modelo a usar
          example: qwen2.5-72b-instruct
        messages:
          type: array
          description: Array de mensajes en formato OpenAI
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant]
                example: user
              content:
                type: string
                example: Hola, ¿cómo estás?
        temperature:
          type: number
          format: float
          minimum: 0
          maximum: 2
          default: 0.7
          description: Temperatura para la generación
        max_tokens:
          type: integer
          default: 2048
          description: Máximo de tokens a generar
        stream:
          type: boolean
          default: false
          description: Si se debe usar streaming
...
    ChatResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        content:
          type: string
          description: Contenido de la respuesta
          example: Hola, estoy bien. ¿En qué puedo ayudarte?
        model:
          type: string
          description: Modelo usado para generar la respuesta
          example: qwen2.5-72b-instruct
        usage:
          type: object
          description: Información de uso de tokens
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
        cached:
          type: boolean
          description: Si la respuesta fue servida desde cache
          example: false
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Código de error estandarizado
              example: ERR_API_KEY_INVALID
            message:
              type: string
              description: Mensaje de error descriptivo
              example: API Key inválida o no autorizada
            statusCode:
              type: integer
              description: Código HTTP del error
              example: 401
            details:
              type: object
              description: Detalles adicionales del error
            retryable:
              type: boolean
              description: Si el error es retryable
              example: false
            timestamp:
              type: string
              format: date-time
              description: Timestamp del error
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para autenticación (opcional para endpoints públicos)
security:
  - ApiKeyAuth: []